完善余额同步，如果同一帐号两次操作时间间隔太短，依然有问题。 只有以后再慢慢优化了。

+++ XZMY.Manage.WindowsService/Service/HyxxService.cs
@@ -123,7 +123,7 @@ namespace XZMY.Manage.WindowsService.Service
             if (dt.Rows.Count == 0)
                 return 0;
 
-            return dt.Rows[0]["knje"].ToString().Trim().ToInt32(0);
+            return dt.Rows[0]["knje"].ToString().Trim().ToDecimal(0);
         }
 
         #endregion
